redux action and reducer files structured
