+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import search from './search';
+import library from './library';
+
+export default combineReducers({
+	search,
+	library
+});
